Time: 4 ms (56.48%), Space: 5.7 MB (95.18%) - LeetHub

+++ 69-sqrtx/69-sqrtx.cpp
@@ -3,12 +3,12 @@ public:
     int mySqrt(int x) {
         if(x == 1 || x == 0)
             return x;
-        long long int s = 0, e = x;
+        long int s = 0, e = x;
         
         while(s < e)
         {
-            long long mid = s+((e-s)>>1);
-            long long int temp = mid*mid;
+            long mid = s+((e-s)>>1);
+            long int temp = mid*mid;
             if(temp == x)
                 return mid;
             else if(temp < x)
